requirements.txt
aiogram==3.15.0
requests==2.32.3
python-dotenv==1.0.1
aiohttp==3.11.11
sqlalchemy==2.0.36
openai>=1.0.0
